Replaced hard coded test data with Faker Replaced all instances of hard-coded property values with randomly generated ones from Faker. Moved all model fixtures to the conftest.py.
