+++ setup.py
@@ -24,6 +24,7 @@ requires = [
 
 tests_require = [
     'WebTest >= 1.3.1',  # py3 compat
+    'Faker',
     'pytest',
     'pytest-cov',
 ]
